requirements.txt
annotated-types==0.7.0
anyio==4.4.0
azure-core==1.30.2
azure-identity==1.16.1
buildozer==1.5.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
docutils==0.21.2
email_validator==2.1.2
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msal==1.28.1
msal-extensions==1.1.0
openai==1.34.0
orjson==3.10.5
packaging==24.1
pexpect==4.9.0
platformdirs==4.2.2
portalocker==2.8.2
ptyprocess==0.7.0
pycparser==2.22
pydantic==2.7.4
pydantic-extra-types==2.8.2
pydantic-settings==2.3.3
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.8.0
pypiwin32==223
python-dotenv==1.0.1
python-multipart==0.0.9
pywin32==306
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
sh==2.0.7
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
virtualenv==20.26.2
watchfiles==0.22.0
websockets==12.0